server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Anatol Piven <anatoly.piven@solbeg.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
alembic = "^1.10.4"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
envparse = "^0.2.0"
python-jose = "^3.3.0"
pytest = "^7.3.1"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^1.10.7"}
python-multipart = "^0.0.6"
httpx = "^0.24.0"
rdkit = "^2023.3.1"
minio = "^7.1.14"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"
isort = "^5.12.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
